fix general info
